Attach the check if the UI should be hidden on mousedown
If there is an input, select, textarea and they are part of the component's hierarchy,
React first will update the elements and then the code which checks should we hide
the UI or not will be invoked. If not, the code will be invoked first and then React will
change the components. The check is: if the element on which user clicked is not a
part of our UI, or it is not the editable area itself - hide the UI. And because React
discards the clicked element before the check to occurs - the UI hides.

To fix that, instead to subscribe to 'click' and do the check, we will subscribe to 'mousedown'.
In this case, we will do the check and then we will trigger the change of the elements of the Toolbar.

Fixes #263