Reset f2f3dc256961928aa613198cfcc71e7b32752ea1
Changing from key down to key up causes the following error:
if navigating with the keyboard on toolbar styles, pressing Enter on
button link causes rendering function of the toolbar to be
called twice. After the second call, the toolbar renders with
the default buttons instead with button-link-edit due to renderExclusive
flag set to false.

Fixes #128