Bump karma-coverage version as recommended by `npm audit`
`npm run test` still clean after this change. Tested that coverage
report is correctly generated by cleaning the output folder (`rm -r
coverage/*`) and then inspecting afterwards:

```
$ tree coverage
coverage
└── Chrome\ 71.0.3578\ (Mac\ OS\ X\ 10.13.6)
    ├── base.css
    ├── index.html
    ├── prettify.css
    ├── prettify.js
    ├── sort-arrow-sprite.png
    └── sorter.js

1 directory, 6 files
```

The coverage report is empty, but that is a separate problem: it is
empty on the current HEAD of the 2.x-develop branch as well
(78486e1747e35647664), and I suspect we're likely to move away from
Karma before getting to fix that issue.