Update webpack-cli to solve build issues
I'm continuing with the security audit for:

    https://issues.liferay.com/browse/IFI-405

and was about to test a new version of webpack-dev-server, but found
that all build commands are currently bombing out due to the use of the
old webpack-cli package. Thanks @julien for pointing me in the right
direction on that one.

So, this commit updates that dep. I confirmed that the build now works.