Port uibridge/richcombo.jsx from createReactClass
This is much like the preceding commits in this PR, but not we had to
move the `getInitialState()` logic into the constructor to avoid this
warning:

```
Warning: getInitialState was defined on %s, a plain JavaScript
class. This is only supported for classes created using
React.createClass. Did you mean to define a state property instead?
```

Test output:

```
 - RichComboBridge :
   * should create a rich combo and invoke its initialization methods : ok
   * should render just the menu button when not expanded : ok
   * should show a dropdown with the action buttons when expanded : 2 skipped
   * should execute the onClick method with the item value when clicking on an item : skipped
```

Related: https://issues.liferay.com/browse/IFI-409