[IFI-411] Add prettier support
For now, just a simple manual invocation (`npm run format`) to fix
things up (`npm run prettier` also offered as a synonym, and `npm run
checkFormat` to report files which aren't curretly formatted). The
package.json tasks match the names that we're using for the same in
Clay.

Note the `.prettierignore` file in "src/lang" to prevent the generated
files from being needlessly formatted. (Although the truth is, we could
format them too if we wanted.)

Once we're happy with this we can look at enforcing correct
formatting with the linter in CI.

Closes: https://issues.liferay.com/browse/IFI-411