Remove require-dir dependency
This dependency doesn't do much for us: removing it shows that we can
get by without it (note how we used four lines invoking the package
before this commit, and when we remove the package we can get an
equivalent outcome with four lines, leaving the total count unchanged).

In any case, the real motivation here is that when we move to Gulp v4
(see #980), the pattern is to refer to tasks using function references
rather than strings, which means we'd have to manually import the
"build" tasks additionally anyway. Reducing our dependency footprint is
just a bonus.

Test plan: `npm run build` and see that everything still works.