Fixes #983 - Make sure code is correctly formatted before publishing
Add a "prepublishOnly" task that errors out if we try to publish code
that isn't formatted with Prettier. Also runs the tests for obvious
reasons.

Once we have linting back up and running we can fail even faster with
respect to formatting and warn or error at lint-time.

This commit includes a fix for one incorrectly formatted file. There are
also a couple of others that need re-formatting:

- scripts/build/gulp/tasks/build.js
- scripts/build/gulp/tasks/css.js

but I am leaving those so as to avoid an unnecessary conflict with #982.