Fix autofixable lint errors
Note that despite the use of the eslint-config-prettier, the actual
sequence here is:

    npm run fix
    npm run format

Because while the config stops ESLint from complaining about things that
Prettier will take care of, ESLint isn't always capable of producing the
right indent from Prettier's point of view.

Before:

  ✖ 1480 problems (1480 errors, 0 warnings)

After:

  ✖ 704 problems (704 errors, 0 warnings)

I'll continue to look at the ones that can't be autofixed.