Add "lint:quiet" to package.json scripts
To make it at least somewhat discoverable. You can already achieve the
same with `npm run lint -- --quiet`, but this makes the availability of
this functionality more obvious to anybody reading the package.json
file.

This is important because we currently have 235 warnings (related to
JSDoc) which, realistically, we won't be able to fix for a long time.

Bonus: moved `--fix` to right after the executable name in the
neighboring "fix" task for consistency.