Silence and fix lints about capital-case non-constructor functions
These are our HOC wrapper functions that all start with capital letters
even though they are not called with `new`. We may remove some (or all)
of these in favor of other abstractions, so rather than change their
names, I am suppressing the lints related to these functions.

Fixed a couple of capitalization inconsistencies along the way as a
result of doing this lint:

- ButtonKeyStroke -> ButtonKeystroke (to match other usages)
- columnResizer -> ColumnResizer (because it really is a constructor)

Before:

    ✖ 688 problems (453 errors, 235 warnings)

After:

    ✖ 572 problems (337 errors, 235 warnings)