Add eslint-plugin-react to get rid of a bunch of errors
ESLint will get confused without this plug-in, for example, thinking
that an import is unused if it appears as a JSX component.

So, we add the plug-in and that takes us from:

    ✖ 572 problems (337 errors, 235 warnings)

to:

    ✖ 305 problems (305 errors, 0 warnings)

Only cheating slightly here by turning "react/prop-types" down from
"error" to "warn". Without this the error count skyrockets and the whole
thing becomes unworkable; making these warnings means that we can chew
through them slowly over time.