Fix lints (and one bug) in src/adapter/main.js
Fixes these lints:

```
   71:43  error  Unnecessary escape character: \/                    no-useless-escape
   75:35  error  Unnecessary escape character: \?                    no-useless-escape
   80:3   error  Expected an object to be thrown                     no-throw-literal
  199:4   error  Unexpected 'this'                                   no-invalid-this
  260:7   error  'regexBasePath' is assigned a value but never used  no-unused-vars
  260:32  error  Unnecessary escape character: \/                    no-useless-escape
```

The bug:

`regexBasePath` was indeed unused, because it wasn't exported in the
refactor done in bdbd6b03df5880f49e8209a846ab388d7982bf51. This is all a
bit magical, but here is how it works:

Our webpack build uses "adapter/main" as an entry point with
configuration `library: 'AlloyEditor'` and `libraryTarget: 'window'`;
this has the effect of exposing everything exported by the adapter as
properties of `window.AlloyEditor` (but note, not `regexBasePath`
because it wasn't exported).

Another webpack build target uses "scripts/build/index" as an entry
point, which in turn imports `*` from the adapter and exposes it as
under `window.AlloyEditor`.

So that's how the exports of the build magically makes the adapter's
exports available on `AlloyEditor`, so that calls like
`AlloyEditor.[something]` can work.

But as I said above, this is *not* happening for `regexBasePath` even
though the adapter itself uses it (`getBasePath` is called by `getUrl`
which is called by `loadLanguageResources`). Why isn't this blowing up
in the release? Two reasons:

1.  In the open source release when we hit the line that does
    `match(AlloyEditor.regexBasePath)`, that's really equivalent to
    `match(undefined)`, which doesn't blow up and instead just matches
    against the beginning of the string. There is more code below this
    that I didn't analyze: maybe it gets through somehow; or maybe it
    blows up when we throw an error ("installation path could not be
    automatically detected"). The error instructs people how to work
    around the error anyway (by setting the `ALLOYEDITOR_BASEPATH`
    global variable), so perhaps that is how people are dealing with it.
2.  In Liferay Portal, we overwrite `AlloyEditor.regexBasePath` at
    runtime because our built bundle has a different name (ie. with a
    "liferay" prefix), and so the default regex wouldn't work for us.

The fix, then, is to export `regexBasePath`.

At this point, src/adapter/main.js is lint-free.