Suppress expensive stats during webpack builds
According to:

    https://webpack.js.org/guides/build-performance/

accessing fields on the stats object can be expensive. So, I changed
things here to only access it if `stats.hasErrors()` is true, or if you
pass in an environment variable:

```
env STATS=1 npm run build
```

By default, stats won't be shown. Instead, you'll see this in the build
output:

```
[15:30:44] Starting 'build:webpack.dev'...

Not showing webpack stats; run with `env STATS=1` to show

[15:31:07] Finished 'build:webpack.dev' after 23 s

Not showing webpack stats; run with `env STATS=1` to show

[15:31:11] Finished 'build:webpack.prod' after 27 s
```

In testing this, I found a bug in the original code which explains why
the stats output is not in color. I was calling `stats.stats.toString({colors:
true})` but I needed to call `stats.toString({colors: true})`. It still
"worked", but not in color.

The bottom-line? Sadly this doesn't trim anything much noticeable off
the build time, but at least it fixes a bug and makes the output
cleaner, so going with it.