Fix lints in plugins/Resizer.es.js
```
  19:24  error  'box' is defined but never used    no-unused-vars
  71:38  error  'index' is defined but never used  no-unused-vars
```

I just removed the `index` param from the `forEach` call because it is
totally standard to omit it.

For the `box` param, I added configuration to mark params starting with
`_` as unused, because all the neighboring methods take a `box` and use
it, and I wanted this one to basically say, "I know all my neighbors
take a `box` and I could use it too but I don't need to".

Related: #990