Deal with 12 remaining "no-used-vars" lints related to function arguments
Fixes:

```
src/components/main.jsx
   50:4   error  'event' is defined but never used. Allowed unused args must match /^_/  no-unused-vars
  228:21  error  'event' is defined but never used. Allowed unused args must match /^_/  no-unused-vars

src/components/toolbars/toolbar-add.jsx
   54:21  error  'prevProps' is defined but never used. Allowed unused args must match /^_/  no-unused-vars
   54:32  error  'prevState' is defined but never used. Allowed unused args must match /^_/  no-unused-vars

src/components/toolbars/toolbar-styles.jsx
   51:21  error  'prevProps' is defined but never used. Allowed unused args must match /^_/  no-unused-vars
   51:32  error  'prevState' is defined but never used. Allowed unused args must match /^_/  no-unused-vars

src/core/table.js
  384:20  error  'editor' is defined but never used. Allowed unused args must match /^_/  no-unused-vars

src/plugins/dragresize_ie11.js
   104:42  error  'event' is defined but never used. Allowed unused args must match /^_/  no-unused-vars

src/plugins/dragresize.js
   44:35  error  'box' is defined but never used. Allowed unused args must match /^_/    no-unused-vars
   95:39  error  'evt' is defined but never used. Allowed unused args must match /^_/    no-unused-vars
  231:5   error  'index' is defined but never used. Allowed unused args must match /^_/  no-unused-vars

src/plugins/embedurl.js
  465:33  error  'event' is defined but never used. Allowed unused args must match /^_/  no-unused-vars
```

For all of these, I indicated that the parameters are intentionally
unused by prefixing theme with an underscore. The one exception was the
`forEach()` call in dragresize.js, which I just removed, because it is
customary to omit the index parameter in that case.

Note that there is a smell here that I am not dealing with: those
`componentDidUpdate()` calls where we are ignoring both the props and
the state are probably a sign that we are doing things in too much of an
imperative way. But refactoring is not the goal here (lint fixing is) so
I am moving on for now.

Test plan: `npm run dev && npm run test && npm run start` and check the
demo.

Related: https://github.com/liferay/alloy-editor/issues/990<Paste>