Fix no-unused-vars lint in button-link-edit-browse.jsx
```
src/components/buttons/button-link-edit-browse.jsx
  67:9   error  'instance' is assigned a value but never used          no-unused-vars
```

There are other lint issues in this file, so this just fixes the safe
one. At least a couple of the others relate to this method because of
the "TODO" noted down the bottom (the method does nothing).

In any case, the move to arrow functions assigned to instance properties
(in 73f21f1647c779a645f67e) makes the `instance` variable unnecessary,
so we remove it and kill off one more lint.

Related: https://github.com/liferay/alloy-editor/issues/990