Don't use process.env.PWD in build
Building on Windows is not officially supported, but we can easily
remove one issue on that platform by removing our use of the PWD
environment variable, which won't be set on Windows.

I can't remember why I did it this way, but I am guessing it was to
avoid the ugliness of the "../../../..".

Test plan: add a `console.log` statement to sanity check the constructed
path and do `npm run build`.