Remove super() calls from ButtonKeystroke HOC
I am looking at teasing apart some of these HOCs. Starting with this one
because it is only used in a few places. This one calls "super" but
looking at the 6 callsites, none of them implement the overridden
methods. So I am going to remove the `super()` calls to make it clear
that there is no interaction with the superclass and these can probably
be replaced with more conventional classes that don't require
inheritance.

For reference, the 6 callsites are:

- src/components/buttons/button-embed.jsx
- src/components/buttons/button-link-browse.jsx
- src/components/buttons/button-link.jsx
- src/components/buttons/button-italic.jsx
- src/components/buttons/button-bold.jsx
- src/components/buttons/button-underline.jsx

And `ButtonKeystroke` is not exported anywhere as far as I can tell, so
we can be confident that there are no unknown callers.

Related: https://github.com/liferay/alloy-editor/issues/1037