Start using React context to share editor instance
Instead of passing the editor through props everywhere, we can get it
out of props and into context. This will make dealing with #1037 easier,
I hope, because it will be one less prop we have to worry about
propagating through our HOCs; but I think it makes sense to do this
anyway because the editor is a "global" value ("global" per editor
instance, that is), which is exactly the use case that context exists
for.