Fix issues with accessing context in constructors
Until React 16.6.0, you couldn't access `this.context` in lifecycle
methods like `componentDidMount()` and friends. With current React you
can, but you still can't do it in constructors, so the recommended
practice is to use a context consumer to pass in the context via a prop:

https://github.com/facebook/react/issues/12397#issuecomment-375501574

I've done an audit of places where we access context from a constructor,
and found these sites that either do it directly or indirectly (via
`getInitialState()` or similar). Testing in the demo, I was able to find
some bugs (where `this.context.editor` didn't work) and see that this
commit fixes all the ones I found.

One thing to note: the `EditorContext.toProps` helper that I added is
yet another HOC -- :sadface: -- but we only need it in a few places, so
I am ok with it. The other thing to note is that when you use it it
should always be the first HOC, because it returns a function component;
if you don't do that (and instead wrap it inside another HOC), then
things will break because our HOCs all work by extending classes.
Extending classes is a brittle way to do things because it limits
flexibility of composition (as seen here); one of the goals in #1045 is
to see if we can refactor our way out of that situation.

Test plan: `npm run dev && npm run test && npm run start` and check All
The Things in the demo with the console open looking for errors.