Remove unnecessary forwarding of editor prop
In previous commits I taught all the components to get the editor from
the context. In this commit I go through and check all the places where
we are passing a prop and remove them after checking that all of their
children are definitely not expecting to receive the editor in a prop.

As such, I actually delete some of the `contextType` code that I added
in the previous intermediate steps, because some of these components end
up being mere passthroughs that accessed the editor for the sole
purposes of forwarding it on. This is why context is a win for this kind
of use case.

Test plan: `npm run dev && npm run test && npm run start` and thoroughly
check demo with the console open watching for errors.