Remove references to required imageTPL prop
This is leading to this in the console on clicking the left-hand-side
"+" button:

```
checkPropTypes.js:19 Warning: Failed prop type: The prop `imageTPL` is marked as required in `ButtonCommand`, but its value is `undefined`.
    in ButtonCommand (created by WidgetArrowBox)
    in WidgetArrowBox (created by WidgetExclusive)
    in div (created by WidgetExclusive)
    in WidgetExclusive;
```

The prop is not used in the component, so it doesn't seem to be required
after all. I also looked in the Liferay Portal and see no other mentions
of tis property name, so I think it is safe to remove.

Test plan: `npm run dev && npm run test && npm run start` and hit the
demo.