Test our debounce implementation instead of CKEDITOR's
If you look in src/core/index.js, you'll see that we were importing
"debounce" from our debounce module and re-exporting it. However, if you
look at our debounce module, you see that we aren't exporting the
function; instead, we're monkey-patching CKEDITOR to supply an
implementation if it doesn't already have one. I don't know under what
circumstances we expect it to have or not have one.

In any case, it seems pointless to test the CKEDITOR implementation when
we should be testing our own. So, we make these changes:

- Export our function from the debounce module, but keep the
  monkey-patching as it was.
- Test our implementation in the tests.