Fix broken language asset updates
Updates were broken in 2b0f4c240, which added a regex which couldn't
possibly match any of the files because none of them are of the form
"a.js". The new regex added in this commit matches "en.js", "en-au.js"
etc. We could make it stricter, but I don't think that's necessary.