Don't pass "this" with "actionPerformed" event from useExecCommand
As noted in the previous commit, there are two places in the
liferay-portal repo where we are breaking encapsulation and reaching
into the component instance to access its `props.commmand` member.
Instead, let's just pass the command directly.

NOTE: If we end up merging this PR, we'll need to land the corresponding
change in the portal at the same time in these two places:

- https://github.com/liferay/liferay-portal/blob/4b5542e86a251faa27c29a8d52cd8a04bde141b7/modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/editor/editor_field.js#L152
- https://github.com/liferay/liferay-portal/blob/399e1b734a97033f4da9f14dbd676ef24019314d/modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/metal/Editor/Editor.es.js#L205