Fix #1109 auto-embedding of pasted video (etc) URLs
The "embed" plugin checks to see whether pasted text matches a regex for
strings starting with "http://" (or "https://"), but the "autolink"
plug-in also listens to paste events and ends up transforming any links
into anchor tags before the "embed" plug-in sees them. This means that
the regex doesn't match and the embed will never work.

In this commit we add a priority mechanism in order to allow the "embed"
plug-in to run first.

The gotcha: when we paste a link, the "selectionChange" handler runs in
the "embed" plug-in and ends up firing an "editorInteraction" event with
the new selection, and that means we see the link toolbar pop-up.
Previously on pasting a link, no toolbar would appear. Still need to
find a solution for that.

Test plan: `npm run dev && npm run test && npm run start`; in demo,
paste normal links, video links, video links with extra query params
etc:

- https://nodesecurity.io/advisories/755 (a normal link)
- https://www.youtube.com/watch?v=s5VpRV9yfuc (a video)
- https://youtu.be/s5VpRV9yfuc?t=178 (same video, but with a starting
  timestamp)

Closes: https://github.com/liferay/alloy-editor/issues/1109