Add useButtonActionStyle hook
Aim is to use this in the following components that currently use the
`ButtonActionStyle` HOC:

- src/components/buttons/button-h1.jsx
- src/components/buttons/button-h2.jsx
- src/components/buttons/button-styles-list-item.jsx
- src/components/buttons/button-code.jsx

Of those, `<ButtonStylesListItem />` is going to be the easiest to
migrate because it doesn't use any other HOCs apart from `ButtonStyle`.
`useButtonActionStyle` composes `useButtonStyle`, which means that we
can replace both HOCs in one go. I'm going to do that in the next
commit, and then tackle the remaining three components (which use
`ButtonStateClasses`, so I think we'll need a hook for that before we
can proceed there).