Add useButtonStateClasses hook and use it in <ButtonCode />
`<ButtonCode />` becomes stateless/functional as a result and quite a
bit simpler.

I am not happy with the names of this and the other hooks, to be honest,
but for now going with a straight swap from the old HOC names (ie. just
appending `use`).

This one barely even needs to be a hook, but doing it anyway to explore
the problem space a bit. I suspect that before the end of all this we
can refactor further to simplify.

Test plan: `npm run dev && npm run test && npm run start`; use code
button in the demo; see visual update and presence of
"ae-button-pressed" class when inspecting in the DOM.