Don't use randomly generated keystroke command names
Doing this to facilitate the migration of the ButtonKeystroke HOC to a
hook-based implementation, but I think we should do this anyway. The
purpose of the randomness here is to avoid command name collisions, but
we can easily do that by prepending a "keystroke:" prefix to each name.
The randomness seems like a bit of a smell because, ideally, you'd have
an utterly deterministic runtime state every time you fire up the
editor (for testability, reproducibility etc). In this case, each
native instance is isolated from others on the page, so I don't think
there's any need for randomness to avoid collisions there either.

As far as hooks go, I want the name to be stable so that I can look up
button keystroke commands by name across mount/unmount events.