refactor: don't use randomly generated keystroke command names (#1314)
The purpose of the randomness here is to avoid command name collisions,
but we can easily do that by prepending a "keystroke:" prefix to each
name. The randomness seems like a bit of a smell because, ideally,
you'd have an utterly deterministic runtime state every time you fire
up the editor (for testability, reproducibility etc). In this case,
each native instance is isolated from others on the page, so I don't
think there's any need for randomness to avoid collisions there either.

Cherry-picking this from the old "research/hooks" branch which we never
merged, seeing as I just audited the ButtonKeystroke stuff for #1314:

https://github.com/liferay/alloy-editor/commits/research/hooks