Update `keystroke.fn` prop in `<ButtonItalic />`
This worked because we overwrite the `fn` prop, but I meant to default
this to an arrow function, not a string, like I did in the previous
commits. Note that when I transformed `ButtonKeystroke` into a hook, I
ripped out the functionality that accepted a string or a function,
because with hooks (and therefore functional components) we don't have
instances with methods that we can look up by sting.