Create `useMergedProps` hook and port `<ButtonLinkBrowse />`
Two components used the `ButtonProps` HOC. This commit creates an
equivalent hook, `useMergedProps`, and uses it to convert
`<ButtonLinkBrowse />` to a stateless functional component.

I'm not crazy about the `useMergedProps` name as it sounds very generic
but this HOC/hook is used in exactly two places, and as you can see, it
has a hard-coded reference to the "linkEdit" config.

One other note: rather than switch from `ButtonStateClasses` to
`useButtonStateClasses`, I dropped it. That HOC adds a
`getStateClasses()` method to the instance, but there are no callers of
it and it is doing nothing.

Some tiny related fixes included in this one along the way because I am
running out of time.

Test plan: AFAIK, this button is only used in the portal, so I build and
tested it there with `console.log` statements to confirm that it was
being invoked as I invoked it with the mouse and with the keyboard.