Improve prop handling in <ButtonEmbed />
This is closer to what we had in the pre-hook implementation where we
just passed along all the props to `<ButtonEmbedEdit />`. Inspecting
the latter shows that it does indeed consult some of these props (eg.
`renderExclusive`, `tabIndex`) plus others that we don't even mention
explicitly here, like `manualSelection` and `cancelExclusive`.