Teach useMergedProps to return a thunk
We can only call hooks at the top level (not inside conditionals), but
we can avoid eagerly performing the merge by returning a thunk instead
of the merged config.

At the same time pass down all props instead of selecting a subset,
matching the change we made in the last commit.