Propagate context in tests
This was the cause of a bunch of failures that are being silently
swallowed (this fixes many failures, but we still need to tackle the
problem that failures are getting swallowed).

Test plan: run Karma in debug mode with `singleRun: false` and look at
failures in the console; see that context-related failures are now gone.