Make utils-ckeditor.js and utils-alloy-editor.js load-order independent
We have a fragile implicit requirement here that utils-ckeditor.js must
be evaluated before utils-alloy-editor.js, which could easily happen by
making an unwitting edit of the karma.common.js file.

This commit changes utils-ckeditor.js so that it can be loaded before or
after. I just changed it to use the same style as utils-alloy-editor.js.
Note that both these files bypass webpack so it's written in an
old-school format.

Note that I also needed to refactor to remove an IIFE in
utils-alloy-editor.js that introduced an order-sensitivity where there
didn't need to be one.

Test plan: `npm run test` with the order of the two files swapped in the
config; swap the order back and run again; in both cases see everything
working. In this commit I am including the swapped order because having
things alphabetical reads better.