Produce a single webpack bundle for use in test suite
As explained in the comment at the head of "test/index.js", added in
this commit, we were creating a huge number of redundant bundles (almost
200) for our test runs, leading the browser to download, parse and
evaluate around 65 MB(!!!) of JavaScript. In addition to being
incredibly wasteful, it also lead to a large number of test failures.

Because the redundant bundles led basically every test to have its own
copy of React, its own copy of EditorContext etc, a large number of
tests that relied on strict equality and sharing module-local state were
completely broken. For example, everything that relied on React Context
(which is almost everything) was broken because each thing would get its
own context created by `React.createContext()`, so consumers wouldn't
match up with providers and no context would get propagated.

This is unfortunately a bit magical and dynamic for my liking, but it is
better than being broken.

Test plan: `npm run test:debug` and see almost all of the pre-existing
failures go away. Will continue hunting down the others in a separate
PR.