Add "cruft" check in utils
In troubleshooting these test failures I have found that poorly behaved
tests may fail to clean-up the DOM, which in turn causes CKEditor to
throw an error when it attempts to attach to a DOM element and already
finds an editor on it.

We should fix those misbehaved tests, but in the name of resilience, I
am adding this check that warns when cruft is found in the DOM, and
cleans it up so that subsequent tests don't fail unnecessarily due to
multiple "#editable" elements being in the DOM. In this way, you see the
test that failed to clean up rather than dozens of innocent tests that
come after it throwing spurious errors.