Unify before/after hooks
There was a lot of scope for human error in the way we had our
before/after hooks set-up. You had to remember which create/destroy hook
to use (AlloyEditor or CKEditor), remember whether to add it to
`beforeEach`/`afterEach` or `before`/`after`, and also remember to
accompany it with a call to `Utils.beforeEach`/`Utils.afterEach`.
Additionally, this all require you to remember to call the function in a
way that sets or preserves the value of `this`, and to pass in a "done"
callback.

To make this better, this commit:

- Folds the contents of `Utils.beforeEach`/`Utils.afterEach` into the
  other utils. ie. `Utils.beforeEach` gets folded into
  `Utils.createAlloyEditor` and into `Utils.createCKEditor`;
  `Utils.afterEach` gets folded into `Utils.destroyAlloyEditor` and
  `Utils.destroyCKEditor`.
- Deletes `Utils.beforeEach` and `Utils.afterEach`, and all calls to
  them.
- Changes all `before`/`after` hooks to use `beforeEach` and `afterEach`
  instead; they are fast enough to run before and after every example,
  and we have been bitten repeatedly by stale test set-up persisting
  across examples and causing spurious failures.
- Moves the `assert.ok` checks up to the top of their respective hooks
  to make it clear that they are guards.
- Adds `assert.ok` calls to enforce that a "done" callback is always
  correctly passed.