Fix button-link-edit-browse tests
These were relatively new tests written in a modern style, but the use
of arrow functions breaks all the (many) places in the test
infrastructure were we rely on `this` being transparently set to the
right value. That's a pretty brittle pattern, but it is too entrenched
to change at this point, so let's just unbreak the test. We would have
noticed this if it were not for failures being silently swallowed:

Related: https://github.com/liferay/alloy-editor/issues/1123