Fix various test issues related to use of React.findDOMNode and friends
`React.findDOMNode` is deprecated and won't work on functional
components anyway, so a lot of these interactions were broken. These
changes fix a number of failures, but there are still many others and it
is not always clear whether any given change has had unintended
side-effects so I am going to keep working on fixing broken tests and
will likely be following this up with more commits.