Move most of scripts/test/*.js into test/lib/*.js
"scripts/test/*.js" doesn't go through webpack by default, which forces
us to write in an old-school style. Move this stuff into "test/lib/"
instead, so that we can be more flexible in our choice of syntax.

"scripts/test/bender.js" is the only thing that remains, and that is
third-party code that we're not going to update anyway, so it can stay
where it is.