Forward refs in EditorContext.toProps
Fixes console warning:

> Warning: Function components cannot be given refs. Attempts to access
> this ref will fail. Did you mean to use React.forwardRef()?`

Which is happening inside of `<ButtonLinkBrowseEdit />` when it renders
`<ButtonLinkEdit />` and attempts to pass a "ref".

Noticed while inspecting console output during `npm run test:debug`. I
think this is the first actual problem that we've found and fixed during
our test-fixing crusade in:

https://github.com/liferay/alloy-editor/issues/1123

We have a line in the component that tries to use the ref like this:

```
const linkTarget = this.linkEditButtonRef.current
        ? this.linkEditButtonRef.current.state.linkTarget
        : '';
```

As far as I can tell, `this.linkEditButtonRef.current` is probably
always null-ish, which means we always default to a linkTarget of ''.
I'm going to check that out in Portal.

Test plan: `npm run build`, `npm run test`, `npm run test:debug`
(inspect console) and `npm run start` (check demo still works). Will
test this in Portal too, which I think is the only place you can
meaningfully test that button.