Don't fire uiReady until component is fully mounted
Follow-up to the parent commit.

It seems that when the parent commit was initially submitted in:

    https://github.com/liferay/alloy-editor/pull/1147

it fixed two tests, but when applied on top of the batch of test fixes
that we're preparing in:

    https://github.com/liferay/alloy-editor/pull/1148

the first of the two tests is still failing.

Changes made here to get it to pass:

- Don't fire "uiReady" until we get the ref from React, at which point
  we know the DOM has been updated and `this._mainUI` is non-null.
- In the tests, move the assertion into the "uiReady" callback, so that
  we don't run it before `_mainUI` has been set.
- Additionally, nest it further inside a `setState` callback, which
  further guarantees that we won't try to assert anything until React
  has flushed the update.
- In the tests, add the `done` callback back and call it only after our
  assertion has run.

Test plan: `npm run build && npm run test`, then `npm run start` and
check demo.