Partial fix for failing toolbar-add.jsx tests
There is some global state here that is not being cleaned up by
`destroyAlloyEditor`. If anything runs before the "toolbar on the right"
test, then the test fails. So move it to the top of the test. That will
fix it only when this file is run alone, which is why I am labelling
this as a partial fix: you have to add `describe.only` to see all these
tests pass. :-(

In the meantime I am going to continue to look into the problem of
global state.